web/botxploit/challenge/requirements.txt
flask 
playwright 
python-dotenv
flask-limiter